Added Try/Except block around header delimiter (#3)
* Added Try/Except block around header delimiter

Headers should end with a carriage return/line feed at the end of each line and then a blank line (with a CRLF) to end the header. In some cases, for example a 406 reponse, the header will not have the blank line and instead the server will close the connection. In this case the code raises a ValueError exception which is not very helpfully. Inserting this try/except block adds clarity to what failed.

* Removed .pyc file, added .gitignore